+++ laktory/models/pipeline/pipelinenode.py
@@ -125,6 +125,10 @@ class PipelineNode(BaseModel, PipelineChild):
         ),
         exclude=True,
     )
+    group: str = Field(
+        None,
+        description="Execution group. Nodes with the same group will be executed together in a single task (if orchestrator supports it).",
+    )
     name: str = Field(..., description="Name given to the node.")
     primary_keys: list[str] = Field(
         None,
@@ -146,6 +150,7 @@ class PipelineNode(BaseModel, PipelineChild):
         None,
         description="Definition of the data sink(s). Set `is_quarantine` to True to store node quarantine DataFrame.",
     )
+    tags: list[str] = Field([], description="Node tags for selective execution.")
     time_column: str | None = Field(
         None,
         description="""